Добавить API архива камер и данных климата

+++ server_tg_home/database/models.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import UTC, datetime
 from typing import Any
 
-from sqlalchemy import DateTime, Float, ForeignKey, Integer, JSON, String, Text
+from sqlalchemy import BigInteger, DateTime, Float, ForeignKey, Integer, JSON, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from server_tg_home.database.session import Base
@@ -39,7 +39,7 @@ class Video(Base):
     job_id: Mapped[str] = mapped_column(ForeignKey("jobs.id"), index=True)
     camera_id: Mapped[str] = mapped_column(String(128), index=True)
     path: Mapped[str] = mapped_column(Text)
-    size_bytes: Mapped[int] = mapped_column(Integer)
+    size_bytes: Mapped[int] = mapped_column(BigInteger)
     duration_sec: Mapped[int | None] = mapped_column(Integer, nullable=True)
     deleted_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=utcnow)
